Refactor documentation and tests
