+++ src/main/java/duke/controller/Evaluator.java
@@ -2,6 +2,8 @@ package duke.controller;
 
 import duke.exceptions.DukeException;
 import duke.tasks.TaskList;
+import duke.ui.AlertBox;
+import duke.ui.DataSaver;
 import duke.ui.Messenger;
 import duke.utils.Parser;
 import duke.utils.Storage;
@@ -24,6 +26,7 @@ public class Evaluator {
      * Evaluates the command passed through the input text area.
      *
      * @param command a string representing the command passed in.
+     * @return a string representing the response from the evaluator.
      */
     public String getResponse(String command) {
         // initialize utilities
@@ -70,6 +73,7 @@ public class Evaluator {
         } catch (Exception e) {
             AlertBox.display("Something went wrong!", e.toString());
         }
+        // error in evaluating
         return "What???";
     }
 }
